Refactoring
Split main.c into different files for message handling and navigation window
